refactor(PopupTriggerText): Extract component to render button text with mobile ext window notice.
