+++ lib/components/narrative/trip-tools.js
@@ -11,17 +11,14 @@ import { Undo } from '@styled-icons/fa-solid/Undo'
 import { withRouter } from 'react-router'
 import bowser from 'bowser'
 import copyToClipboard from 'copy-to-clipboard'
-import coreUtils from '@opentripplanner/core-utils'
 import PropTypes from 'prop-types'
 import React, { Component, useContext, useMemo } from 'react'
 
 import * as uiActions from '../../actions/ui'
 import { ComponentContext } from '../../util/contexts'
 import { IconWithText } from '../util/styledIcon'
-import InvisibleA11yLabel from '../util/invisible-a11y-label'
+import PopupTriggerText from '../app/popup-trigger-text'
 import startOver from '../util/start-over'
-
-const isMobile = coreUtils.ui.isMobile()
 
 // Copy URL Button
 
@@ -238,21 +235,7 @@ const TripTools = ({
             <LinkButton
               Icon={PopupIcon}
               onClick={() => setPopupContent(popupTarget)}
-              text={
-                <>
-                  <FormattedMessage
-                    defaultMessage={intl.formatMessage({
-                      id: `config.popups.${popupTarget}`
-                    })}
-                    id={`config.popups.${popupTarget}-narrow`}
-                  />
-                  {isMobile && (
-                    <InvisibleA11yLabel>
-                      <FormattedMessage id="common.linkOpensNewWindow" />
-                    </InvisibleA11yLabel>
-                  )}
-                </>
-              }
+              text={<PopupTriggerText compact popupTarget={popupTarget} />}
             />
           )
         }
